+++ HighlightEmptyMachines/HighlightEmptyMachines/ModEntry.cs
@@ -1,4 +1,6 @@
-﻿using AtraShared.ConstantsAndEnums;
+﻿using System.Diagnostics;
+
+using AtraShared.ConstantsAndEnums;
 using AtraShared.Integrations;
 using AtraShared.MigrationManager;
 using AtraShared.Utils.Extensions;